Eusuite&DeploymentNando/apps/eusuite-public-backend/requirements.txt
# EUSuite Public Backend - Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
aiosmtplib==3.0.1
jinja2==3.1.3
stripe==7.10.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.6
